Experimental feature: Disable excepting on "no records match"
I've encountered a server where a harvest progressed normally, but then returned a "No Records Match" exception, killing the harvest. I'm suspicious that this might actually be how this server is configured to indicate the end of a harvest.

So, this is an experimental feature to permit treating this error as a "stop" signal rather than a true exception.